// app/types/AdminBondType.ts
export type AdminBondType = {
    country: string;
    description : string;
    differed : number;
    due_date : string;
    entity: string;
    facial_rate: number;
    guarantee: string;
    id: string;
    isin: string;
    period: string
    refund: string;
    space: string;
    total_number: number;  
    type: string;
}